Add blog to context of comment_create_view()
An error was occurring because comment_reaction_section.html was not
getting access to the blog and blog_slug of the particular comment added
to blog_detail.html when a comment is added. So the blog item had to be
added to the context of comment_create_view() for
comment_reaction_section.html to get access to the the blog.